Rename directory name.
